requirements.txt
Django==5.0.7
djangorestframework==3.14.0
django-cors-headers==4.9.0
Pillow==11.3.0
python-decouple==3.8
